use a single VideoElementMock definition
