+++ test/unit/mocks/VideoElementMock.js
@@ -1,3 +1,10 @@
+class TextTrackMock {
+    constructor() {
+        this.kind = null;
+        this.label = null;
+    }
+}
+
 class VideoElementMock {
 
     setup() {
@@ -5,10 +12,20 @@ class VideoElementMock {
         this.volume = 0;
         this.currentTime = 0;
         this.duration = 0;
+        this.textTracks = [];
     }
 
     constructor() {
         this.setup();
+    }   
+
+    addTextTrack(kind, label) {
+        let textTrack = new TextTrackMock();
+        textTrack.kind = kind;
+        textTrack.label = label;
+        this.textTracks.push(textTrack);
+
+        return textTrack;
     }
 
     reset() {
